requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
blinker==1.9.0
blis==1.3.3
cachetools==7.0.5
catalogue==2.0.10
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.13
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.25.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
gpt4all==2.8.2
groq==1.1.1
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.15
narwhals==2.18.0
networkx==3.6.1
numpy==2.4.3
openai==2.26.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
preshed==3.0.12
protobuf==6.33.5
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
sentencepiece==0.2.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
smmap==5.0.3
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.10
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.5
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.3
wrapt==2.1.2
